Tests/LibWeb/Text/input/geometry/domrect-create.html

<script src="../include.js"></script>
<script>
    test(() => {
        let testCounter = 1;
        function testPart(part) {
            try {
                println(`${testCounter}. ${JSON.stringify(part())}`);
            } catch (e) {
                println(`${testCounter}. Exception: ${e.name}`);
            }
            testCounter++;
        }

        for (const Rect of [DOMRect, DOMRectReadOnly]) {
            println(`Testing ${Rect.name}:`);
            
            // 1. Creating a DOMRect with no arguments
            testPart(() => new Rect());

            // 2. Creating a DOMRect with positive x, y, width and height
            testPart(() => new Rect(10, 20, 30, 40));

            // 3. Creating a DOMRect with negative x and y values
            testPart(() => new Rect(-10, -20, 30, 40));

            // 4. Creating a DOMRect with DOMRect.fromRect()
            testPart(() => Rect.fromRect({ x: 10, y: 20, width: 30, height: 40 }));

            testCounter = 1;
        }
    });
</script>
